WIP Implementation of incremental soundtrack
The code for the incremental soundtrack has been put in place and tested.
Before merging this I need to:

-)Have the new 9 tracks of the incremental soundtrack.
-)Assign them to the list of tracks.

We start at 0 memory with no soundtrack at all (I'd like the player to understand it is incrementing while depositing memories).
We do not need track 10 since at ten the music stops and the piano appears.